chore: integrated shared to androidMain to get planner resource

+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/planner/model/From_.kt
@@ -9,7 +9,7 @@ data class From_
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val departure: Int? = null,
+    val departure: Long? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
