+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/planner/model/Plan.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Plan
 (
-    val date: Int? = null,
+    val date: Long? = null,
     val from: From? = null,
     val to: To? = null,
     val itineraries: List<Itinerary>? = null,
