+++ shared/src/commonMain/kotlin/edu/usf/cutr/otp/planner/model/To_.kt
@@ -8,7 +8,7 @@ data class To_(
     val name: String? = null,
     val lon: Double? = null,
     val lat: Double? = null,
-    val arrival: Int? = null,
+    val arrival: Long? = null,
     val vertexType: String? = null,
     val additionalProperties: MutableMap<String,@Polymorphic Any> = HashMap()
 )
